Rename dispatcher to exporter
