+++ agent/agent-exporter-brpc/README.md
@@ -1,0 +1,4 @@
+This module provides an implementation of agent exporter over the BRPC protocol,
+which is a protobuf and JSON based RPC protocol.
+
+See [brpc](../../component/component-brpc/) to learn more.
