--- agent/agent-observability/src/main/java/org/bithon/agent/observability/dispatcher/channel/IMessageChannel.java
@@ -1,29 +1,0 @@
-/*
- *    Copyright 2020 bithon.org
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.bithon.agent.observability.dispatcher.channel;
-
-/**
- * @author frank.chen021@outlook.com
- * @date 2020/10/27 2:17 下午
- */
-public interface IMessageChannel extends AutoCloseable {
-
-    /**
-     * In the current design, this method is called in one thread only, so it's thread-safe.
-     */
-    void sendMessage(Object message);
-}
