+++ agent/agent-observability/src/main/java/org/bithon/agent/observability/jvm/JvmMonitorService.java
@@ -16,8 +16,8 @@
 
 package org.bithon.agent.observability.jvm;
 
-import org.bithon.agent.observability.dispatcher.Dispatcher;
-import org.bithon.agent.observability.dispatcher.Dispatchers;
+import org.bithon.agent.observability.exporter.Exporter;
+import org.bithon.agent.observability.exporter.Exporters;
 import org.bithon.agent.observability.metric.collector.jvm.JvmEventMessageBuilder;
 import org.bithon.agent.observability.metric.collector.jvm.JvmMetricCollector;
 import org.bithon.agent.starter.IAgentService;
@@ -39,15 +39,15 @@ public class JvmMonitorService implements IAgentService {
         //
         // dispatch started message once the dispatcher is ready
         //
-        Dispatchers.getOrCreate(Dispatchers.DISPATCHER_NAME_EVENT)
-                   .onReady((dispatcher) -> dispatcher.send(dispatcher.getMessageConverter()
+        Exporters.getOrCreate(Exporters.DISPATCHER_NAME_EVENT)
+                 .onReady((dispatcher) -> dispatcher.send(dispatcher.getMessageConverter()
                                                                       .from(JvmEventMessageBuilder.buildJvmStartedEventMessage())));
     }
 
     @Override
     public void stop() {
         // dispatch jvm stopped message
-        Dispatcher dispatcher = Dispatchers.getOrCreate(Dispatchers.DISPATCHER_NAME_EVENT);
-        dispatcher.send(dispatcher.getMessageConverter().from(JvmEventMessageBuilder.buildStoppedEventMessage()));
+        Exporter exporter = Exporters.getOrCreate(Exporters.DISPATCHER_NAME_EVENT);
+        exporter.send(exporter.getMessageConverter().from(JvmEventMessageBuilder.buildStoppedEventMessage()));
     }
 }
