+++ agent/agent-observability/src/main/java/org/bithon/agent/observability/metric/collector/MetricRegistry.java
@@ -17,7 +17,7 @@
 package org.bithon.agent.observability.metric.collector;
 
 import org.bithon.agent.instrumentation.expt.AgentException;
-import org.bithon.agent.observability.dispatcher.IMessageConverter;
+import org.bithon.agent.observability.exporter.IMessageConverter;
 import org.bithon.agent.observability.metric.model.IMetricSet;
 import org.bithon.agent.observability.metric.model.IMetricValueProvider;
 import org.bithon.agent.observability.metric.model.schema.Schema2;
