+++ agent/agent-observability/src/main/java/org/bithon/agent/observability/metric/collector/jvm/JvmMetricCollector.java
@@ -16,7 +16,7 @@
 
 package org.bithon.agent.observability.metric.collector.jvm;
 
-import org.bithon.agent.observability.dispatcher.IMessageConverter;
+import org.bithon.agent.observability.exporter.IMessageConverter;
 import org.bithon.agent.observability.metric.collector.IMetricCollector;
 import org.bithon.agent.observability.metric.collector.MetricCollectorManager;
 import org.bithon.agent.observability.metric.collector.MetricRegistryFactory;
