+++ agent/agent-observability/src/main/resources/META-INF/services/org.bithon.agent.starter.IAgentShutdownListener
@@ -1,1 +1,1 @@
-org.bithon.agent.observability.dispatcher.DispatchShutdown
+org.bithon.agent.observability.exporter.ExporterShutdown
