+++ agent/agent-plugins/bithon-sdk/src/main/java/org/bithon/agent/plugin/bithon/sdk/interceptor/MetricRegistryFactory$Create.java
@@ -18,7 +18,7 @@ package org.bithon.agent.plugin.bithon.sdk.interceptor;
 
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.ReplaceInterceptor;
 import org.bithon.agent.instrumentation.loader.PluginClassLoader;
-import org.bithon.agent.observability.dispatcher.IMessageConverter;
+import org.bithon.agent.observability.exporter.IMessageConverter;
 import org.bithon.agent.observability.metric.collector.IMetricCollector2;
 import org.bithon.agent.observability.metric.collector.IMetricCollectorBase;
 import org.bithon.agent.observability.metric.collector.MetricCollectorManager;
