+++ agent/agent-plugins/bithon-sdk/src/main/java/org/bithon/agent/plugin/bithon/sdk/interceptor/MetricsRegistryDelegate.java
@@ -16,7 +16,7 @@
 
 package org.bithon.agent.plugin.bithon.sdk.interceptor;
 
-import org.bithon.agent.observability.dispatcher.IMessageConverter;
+import org.bithon.agent.observability.exporter.IMessageConverter;
 import org.bithon.agent.observability.metric.collector.IMeasurement;
 import org.bithon.agent.observability.metric.collector.IMetricCollector2;
 import org.bithon.agent.observability.metric.model.schema.IDimensionSpec;
